fix: remove extraneous border line
Area 5 would have a bogus left-border line on large screen sizes.

As far as I can tell, this is because it tried to reset the border
that smaller-screen-size css would set (border-left: $border-1)
with a 'border: none'. But the border-left instruction would still
win css precedence rules because it an ltr/rtl direction defined.

Changing the 'border: none' to a 'border-left: none', explicitly
only undoing the smaller-screen-size css, seems to fix this
extra line.